requirements.txt
# --- Core Framework ---
fastapi==0.104.1
uvicorn[standard]==0.24.0

# --- Environment Variables ---
python-dotenv==1.0.0

# --- Database ---
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
passlib[bcrypt]==1.7.4

# --- Data Validation ---
pydantic==2.5.0
pydantic-settings==2.1.0

# --- OpenAI Integration ---
openai==1.3.7

# --- HTTP Client (Async) ---
httpx==0.25.2
aiohttp==3.9.1

# --- Testing ---
pytest==7.4.3
pytest-asyncio==0.21.1

